obs/Backend/requirements.txt
fastapi
uvicorn[standard]
obs-websocket-py
python-dotenv